% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbox_map.R
\name{bbox_map}
\alias{bbox_map}
\title{Plot a stamenmap from bbox}
\usage{
bbox_map(bbox)
}
\arguments{
\item{bbox}{Boundaries box generated by \link{get_bbox_from_gpx_table} or
any other Boundaries box tool}
}
\description{
Plot a stamenmap from bbox
}
