% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance.R
\name{calc_distance}
\alias{calc_distance}
\title{Calculate the distance between two places in m}
\usage{
calc_distance(points_from = data.frame(lng = c(), lat = c()),
  points_to = data.frame(lng = c(), lat = c()))
}
\arguments{
\item{points_from}{\code{data.frame} A data frame with a 'lng' at 'lat' column
containing a list of points from where the distance shall be calculated}

\item{points_to}{\code{data.frame} A data frame with a 'lng' at 'lat' column
with just a single point that the distance shall be calculated to}
}
\value{
A table of the distances inside the 'nn.dists' column of the table for each point of the \code{points_from} input
}
\description{
Calculate the distance between two places in m
}
\details{
Can be found at https://stackoverflow.com/questions/39454249/checking-whether-coordinates-fall-within-a-given-radius?noredirect=1&lq=1
}
\author{
Sebastian Wolf \email{sebastian@mail-wolf.de}
}
