% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{el_mat}
\alias{el_mat}
\title{Elevation matrix around Alpe d Huez}
\format{A data frame with 125 rows and 190 variables:}
\source{
\url{https://srtm.csi.cgiar.org/srtmdata/}
}
\usage{
el_mat
}
\description{
data frame containing the elevation data for a sector around
Alpe d Huez with 250 m2 accuracy. Longitudinal degrees are
stored in the last column, latitutes are stored as column names.
}
\keyword{datasets}
