% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{el_mat_eudem}
\alias{el_mat_eudem}
\title{Elevation matrix around Alpe d Huez from EUDEM}
\format{A data frame with 432 rows and 521 variables:}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1.1/view}
}
\usage{
el_mat_eudem
}
\description{
data frame containing the elevation data for a sector around
Alpe d Huez with 30 m2 accuracy. Longitudinal degrees are
stored in the last column, latitutes are stored as column names.
Data derived from EUDEM project.
}
\keyword{datasets}
