% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elevation_data_Schliersee}
\alias{elevation_data_Schliersee}
\title{Elevation matrix around Schliersee}
\format{A data frame with 921 rows and 561 variables:}
\source{
\url{https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1.1/view}
}
\usage{
elevation_data_Schliersee
}
\description{
data frame containing the elevation data for a sector around
Schliersee and Tegernsee in Germany with 30m2 accuracy. Longitudinal degrees are
stored in the last column, latitutes are stored as column names and in the last row.
}
\keyword{datasets}
