% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_arcis_map_image.R
\name{get_arcgis_map_image}
\alias{get_arcgis_map_image}
\title{Download a map image from the ArcGIS REST API}
\usage{
get_arcgis_map_image(bbox, map_type = "World_Street_Map", file = NULL,
  width = NULL, height = NULL, sr_bbox = 4326, type = "MapServer")
}
\arguments{
\item{bbox}{bounding box coordinates (list of 2 points with long/lat values)}

\item{map_type}{map type to download - options are World_Street_Map, World_Imagery, World_Topo_Map}

\item{file}{file path to save to. Default is NULL, which will create a temp file.}

\item{width}{image width (pixels)}

\item{height}{image height (pixels)}

\item{sr_bbox}{Spatial Reference code for bounding box}

\item{type}{MapServer will get a map, else elevation data gets downloaded}
}
\value{
file path for the downloaded .png map image
}
\description{
Download a map image from the ArcGIS REST API
}
\details{
This function uses the ArcGIS REST API, specifically the
"Execute Web Map Task" task. You can find links below to a web UI for this
rest endpoint and API documentation.

Web UI: https://utility.arcgisonline.com/arcgis/rest/services/Utilities/PrintingTools/GPServer/Export%20Web%20Map%20Task/execute
API docs: https://developers.arcgis.com/rest/services-reference/export-web-map-task.htm
}
\examples{
bbox <- list(
  p1 = list(long = -122.522, lat = 37.707),
  p2 = list(long = -122.354, lat = 37.84)
)
overlay_file <- get_arcgis_map_image(bbox)
}
