% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bbox_from_gpx_table.R
\name{get_bbox_from_gpx_table}
\alias{get_bbox_from_gpx_table}
\title{Get bbox format from gpx file table}
\usage{
get_bbox_from_gpx_table(gpx_table, arcgis = FALSE)
}
\arguments{
\item{gpx_table}{data.frame generated with \link{get_table_from_gpx}}

\item{arcgis}{(\code{logical}) Whether a format to download
data from arcgis by \link{get_arcgis_map_image} shall be provided}
}
\value{
list of coordinates box
}
\description{
Get bbox format from gpx file table
}
