% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_distance_from_gpx_table.R
\name{get_distance_from_gpx_table}
\alias{get_distance_from_gpx_table}
\title{Calculate distance between all points}
\usage{
get_distance_from_gpx_table(gpx_table)
}
\arguments{
\item{gpx_table}{\code{data.frame} created by \link{get_table_from_gpx}}
}
\value{
vector of distances in m
}
\description{
This function uses GPS data and \link{calc_distance} to calculate
  the distance between all points
}
