% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevdata_from_bbox.R
\name{get_elevdata_from_bbox}
\alias{get_elevdata_from_bbox}
\title{Generate an elevation matrix from boundary box}
\usage{
get_elevdata_from_bbox(bbox)
}
\arguments{
\item{bbox}{Boundaries box as generated by \link{get_bbox_from_gpx_table}}
}
\value{
A matrix with the latitude in columns and the longitude in rows and the elevation
  as values (in m)
}
\description{
Generate an elevation matrix from boundary box
}
