% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevdata_list.R
\name{get_elevdata_list}
\alias{get_elevdata_list}
\title{Get elevation data labeled for rayshader}
\usage{
get_elevdata_list(elevdata)
}
\arguments{
\item{elevdata}{(\code{data.frame}) generated by get_elevdata_from_bbox}
}
\value{
\code{list} with three elements: \code{elevation_matrix} containing
  the elevation data with the 0.0 coordinate on top right.
  \code{lon} The longitudinal labels of this elevation matrix
  \code{lat} The latitudinal labels of this elevation matrix
}
\description{
Due to the transposing of elevation data at the Greenich line for all my GPX
animations data needs to be transposed. Therefore the labels of longitudinal
and latitudinal also need to be transposed to suit rayshader package
}
