% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevdata_long.R
\name{get_elevdata_long}
\alias{get_elevdata_long}
\title{Derive a long format of elevation data for ggplot}
\usage{
get_elevdata_long(elevdata)
}
\arguments{
\item{elevdata}{(\code{data.frame}) Elevation data build by \link{get_elevdata_from_bbox}}
}
\description{
Derive a long format of elevation data for ggplot
}
