% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_enrichted_gpx_table.R
\name{get_enriched_gpx_table}
\alias{get_enriched_gpx_table}
\title{Enrich gpx data with time and speed}
\usage{
get_enriched_gpx_table(gpx_table, every_x_value = 5)
}
\arguments{
\item{gpx_table}{(\code{data.frame}) table generated with \link{get_table_from_gpx}}

\item{every_x_value}{(\code{numeric}) Just take every xth value to make}
}
\value{
the gpx data with true time and the relative speed between points
}
\description{
Enrich gpx data with time and speed
}
