% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_overlay.R
\name{get_image_overlay}
\alias{get_image_overlay}
\title{Derive an overlay image from ArcGis}
\usage{
get_image_overlay(bbox_arcgis)
}
\arguments{
\item{bbox_arcgis}{\code{list} created with \code{get_bbox_from_gpx_table(..., arcgis = TRUE)}}
}
\value{
An image rendered with \code{png::readPNG} which was downloaded from ArcGis mapserver
}
\description{
Derive an overlay image from ArcGis
}
