% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_from_gpx.R
\name{get_table_from_gpx}
\alias{get_table_from_gpx}
\title{Generate a table from GPX file}
\usage{
get_table_from_gpx(gpx_file_loc)
}
\arguments{
\item{gpx_file_loc}{(\code{character}) location of the gpx file}
}
\value{
A table with the gpx coordinates, elevation and time data
}
\description{
Generate a table from GPX file
}
