% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{places_alpe_de_huez}
\alias{places_alpe_de_huez}
\title{Places labeled around Alpe d Huez}
\format{A data frame with 4 rows and 4 variables:}
\usage{
places_alpe_de_huez
}
\description{
data frame containing the places that should be labeled
during a video rendering around the Schliersee Tour. The
data contains longitude und latitude. Moreover each
place has a title and a label.
}
\keyword{datasets}
