% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2d_animation.R
\name{plot_2d_animation}
\alias{plot_2d_animation}
\title{Animate GPX line on 2d plot}
\usage{
plot_2d_animation(gpx_table, elevdata_long)
}
\arguments{
\item{gpx_table}{(\code{data.frame}) gpx table created by \link{get_enriched_gpx_table}}

\item{elevdata_long}{(\code{data.frame}) elevation data created by \link{get_elevdata_long}}
}
\value{
A gganimate animated plot where the gpx data gets animated on top of an elevation map
}
\description{
Animate GPX line on 2d plot
}
