% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_2d_elevdata.R
\name{plot_2d_elevdata}
\alias{plot_2d_elevdata}
\title{Plot a 2D raster of the bbox}
\usage{
plot_2d_elevdata(elevdata_rayshade, overlay_img = NULL)
}
\arguments{
\item{elevdata_rayshade}{(\code{matrix}) with elevation data but unlabeled
by \code{unlabel_elevdata}}

\item{overlay_img}{(\code{character}) Link to an image to put on-top
of the map. Best created with \link{get_arcgis_map_image}}
}
\description{
Plot a 2D raster of the bbox
}
