% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlabel_elevdata.R
\name{unlabel_elevdata}
\alias{unlabel_elevdata}
\title{Unlabel elevation data matrix for rayshader}
\usage{
unlabel_elevdata(elevdata = NULL)
}
\arguments{
\item{elevdata}{data.frame created by \link{get_elevdata_from_bbox}}
}
\value{
The data.frame without labels in last column
}
\description{
Unlabel elevation data matrix for rayshader
}
