% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_animation.R
\name{video_animation}
\alias{video_animation}
\title{Create a 3D video with an animated GPX route}
\usage{
video_animation(gpx_table = NULL, elevdata_long = NULL,
  number_of_screens = 10, make_gif = TRUE, theta = -10, zoom = 0.7,
  ffmpeg_path = "C:/Programme_2/ffmpeg/bin/", output_file_loc,
  overwrite = TRUE)
}
\arguments{
\item{gpx_table}{(\code{data.frame}) gpx table created by \link{get_enriched_gpx_table}}

\item{elevdata_long}{(\code{data.frame}) elevation data created by \link{get_elevdata_long}}

\item{number_of_screens}{(\code{numeric}) A guess of the number of screens/frames that
shall be shown. An amount of images of the 3D map will be created that suits
this. The more screens there are the more fluent the animation is. Each
screen though takes about 20-30 seconds to render on a decent PC}

\item{make_gif}{(\code{logical}) Whether the screens shall be rendered out as a gif file. Just
works for less than 24 screens}

\item{theta}{Rotation around z-axis.}

\item{zoom}{Zoom factor.}

\item{ffmpeg_path}{(\code{character}) to render more than 24 screens ffmpeg is needed to be in the path. Therefore
please install it and provide the link to the path here. The ffmpeg already in path
will not be overwritten}

\item{output_file_loc}{(\code{character}) Where to store the output mp4 file.}

\item{overwrite}{(\code{logical}) Whether to overwrite the output fie}
}
\description{
Create a 3D video with an animated GPX route
}
