% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_animation_rayshade.R
\name{video_animation_rayshade}
\alias{video_animation_rayshade}
\title{Create a Rayshader animation of a GPX table}
\usage{
video_animation_rayshade(gpx_table, elevdata, bbox_arcgis = NULL,
  number_of_screens = 500, title = "Cycling Trip", places = NULL,
  output_file_loc = tempfile(fileext = ".mp4"), width = 1200,
  height = 800, flyin = list(theta = c(280, 250), zoom = c(0.6, 1.5),
  phi = c(58, 33), duration = 4), overwrite = TRUE)
}
\arguments{
\item{gpx_table}{\code{data.frame} gpx table created by get_enriched_gpx_table}

\item{elevdata}{\code{data.frame} elevation data reveiced by \link{get_elevdata_from_bbox}}

\item{bbox_arcgis}{\code{list} Boundary Box to derive boundaries of an area derived by
\code{get_bbox_from_gpx_table(..., arcgis = TRUE)}}

\item{number_of_screens}{\code{numeric} Number of frames the animation should have}

\item{title}{\code{character} The introductory title shown in an infly video}

\item{places}{\code{data.frame} containing some places where large labels should be
added during the video. needs to have the columns \code{lon, lat, label, title}}

\item{output_file_loc}{\code{character} Where to store the output mp4 file.}

\item{width}{\code{numeric} Width of the screen to render}

\item{height}{\code{numeric} height of the 3d window}

\item{flyin}{\code{list} parameters of \link{video_util_infly}}

\item{overwrite}{\code{logical} Whether the final video file can be overwritten}
}
\description{
Create a Rayshader animation of a GPX table
}
