% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video_util_infly.R
\name{video_util_infly}
\alias{video_util_infly}
\title{Function to move along the rayshader image at first}
\usage{
video_util_infly(theta = c(280, 250), zoom = c(0.6, 1.5), phi = c(58,
  33), title = "Random Bike Ride", width = 1200, height = 200,
  duration = 4)
}
\arguments{
\item{theta}{\code{numeric} Theta angle will move from first value to first value minus second value}

\item{zoom}{\code{numeric} zoom angle will move from second value to first value}

\item{phi}{\code{numeric} phi angle will move from first value to first value minus second value}

\item{title}{\code{character} Title of the infly video}

\item{width}{\code{numeric} width of the inflight video}

\item{height}{\code{numeric} height of the inflight video title}

\item{duration}{\code{numeric} duration in seconds}
}
\value{
\code{character} vector with the absolute paths of the screen images created for the infly
}
\description{
Function to move along the rayshader image at first
}
