% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_image_size.R
\name{define_image_size}
\alias{define_image_size}
\title{Define image size variables from the given bounding box coordinates.}
\usage{
define_image_size(bbox, major_dim = 400)
}
\arguments{
\item{bbox}{bounding box coordinates (list of 2 points with long/lat values)}

\item{major_dim}{major image dimension, in pixels.
Default is 400 (meaning larger dimension will be 400 pixels)}
}
\value{
list with items "width", "height", and "size" (string of format "<width>,<height>")
}
\description{
Define image size variables from the given bounding box coordinates.
}
\examples{
bbox <- list(
  p1 = list(long = -122.522, lat = 37.707),
  p2 = list(long = -122.354, lat = 37.84)
)
image_size <- define_image_size(bbox, 600)

}
