% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_elevdata_from_bbox.R
\name{get_elevdata_from_bbox}
\alias{get_elevdata_from_bbox}
\title{Generate an elevation matrix from boundary box}
\usage{
get_elevdata_from_bbox(bbox, type = c("SRTM", "EUDEM"), file = NULL)
}
\arguments{
\item{bbox}{Boundaries box as generated by \link{get_bbox_from_gpx_table}}

\item{type}{Whether SRTM or EUDEM data should be used for elevation profiles. For EUDEM data
you need to download the right tif file at: https://land.copernicus.eu/imagery-in-situ/eu-dem/eu-dem-v1.1}

\item{file}{In case of EUDEM data provide a link to the GeoTIFF file}
}
\value{
A matrix with the latitude in columns and the longitude in rows and the elevation
  as values (in m)
}
\description{
Generate an elevation matrix from boundary box
}
